// frontend/lib/api/generated/social-security-registration/social-security-registration.ts
export type getApiSocialSecurityRegistrationResponse200TextPlain = {
  data: SocialSecurityRegistrationPagedResult
  status: 200
}

export type getApiSocialSecurityRegistrationResponse200ApplicationJson = {
  data: SocialSecurityRegistrationPagedResult
  status: 200
}

export type getApiSocialSecurityRegistrationResponse200TextJson = {
  data: SocialSecurityRegistrationPagedResult
  status: 200
}
// ...
export type getApiSocialSecurityRegistrationResponseSuccess = (getApiSocialSecurityRegistrationResponse200TextPlain | getApiSocialSecurityRegistrationResponse200ApplicationJson | getApiSocialSecurityRegistrationResponse200TextJson) & {
  headers: Headers;
};