// frontend/lib/api/generated/model/socialSecurityRegistrationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Prevly.Api
 * OpenAPI spec version: 1.0
 */

export type SocialSecurityRegistrationStatus = typeof SocialSecurityRegistrationStatus[keyof typeof SocialSecurityRegistrationStatus];


export const SocialSecurityRegistrationStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const;